"Illinois" (UIUC) Metropolis themed HTML Template for RMarkdown

Generates from an RMarkdown file an HTML presentation with "Illinois" (UIUC) colors and identity standards based on the Metropolis theme.

html_imetropolis(
  self_contained = FALSE,
  seal = TRUE,
  chakra = "https://remarkjs.com/downloads/remark-latest.min.js",
  nature = list(),
  ...
)

Arguments

self_contained
Whether to produce a self-contained HTML file by embedding all external resources into the HTML file. See the ‘Note’ section below.

seal
Whether to generate a title slide automatically using the YAML metadata of the R Markdown document (if FALSE, you should write the title slide by yourself).

chakra
A path to the remark.js library (can be either local or remote). Please note that if you use the default remote latest version of remark.js, your slides will not work when you do not have Internet access. They might also be broken after a newer version of remark.js is released. If these issues concern you, you should download remark.js locally (e.g., via summon_remark()), and use the local version instead.

nature
(Nature transformation) A list of configurations to be passed to remark.create(), e.g. list(ratio = '16:9', navigation = list(click = TRUE)); see https://github.com/gnab/remark/wiki/Configuration. Besides the options provided by remark.js, you can also set autoplay to a number (the number of milliseconds) so the slides will be played every autoplay milliseconds; alternatively, autoplay can be a list of the form list(interval = N, loop = TRUE), so the slides will go to the next page every N milliseconds, and optionally go back to the first page to restart the play when loop = TRUE. You can also set countdown to a number (the number of milliseconds) to include a countdown timer on each slide. If using autoplay, you can optionally set countdown to TRUE to include a countdown equal to autoplay. To alter the set of classes applied to the title slide, you can optionally set titleSlideClass to a vector of classes; the default is c("center", "middle", "inverse").

...
For tsukuyomi(), arguments passed to moon_reader(); for moon_reader(), arguments passed to rmarkdown::html_document().

Value

A modified xaringan::moon_reader() based on the "Illinois" (UIUC) Metropolis themed HTML template.

Author

James Joseph Balamuta (Light Modifications of Metropolis and R function wrapper) See AUTHORS for more specific details behind each of the contributions.

Examples

if (FALSE) {
# Generate slide deck from beamer template
rmarkdown::draft("slide_deck.Rmd", template = "html_imetropolis", package = "uiucthemes")

# Compile the document
rmarkdown::render("slide_deck/slide_deck.Rmd")
}