UIUC Class Journal Format

Modified version of the Journal of Statistical Software (JSS) Format for creating a Journal of Statistical Software (JSS) articles. Adapted from http://www.jstatsoft.org/about/submissions.

latex_journal_format(..., keep_tex = TRUE, citation_package = "natbib")

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use default if neither package is to be used, which means citations will be processed via the command pandoc-citeproc.

Value

R Markdown output format to pass to render

Details

This format is largely based off of the rticles of jss_article.

References

rticles package by JJ Allaire, R Foundation, Hadley Wickham, Journal of Statistical Software, Yihui Xie, Ramnath Vaidyanathan, Association for Computing Machinery, Carl Boettiger, Elsevier, Karl Broman, Kirill Mueller, Bastiaan Quast, Randall Pruim, Ben Marwick, Charlotte Wickham, Oliver Keyes and Miao Yu (2017). rticles: Article Formats for R Markdown. R package version 0.4.1. https://CRAN.R-project.org/package=rticles

Author

JSS Editorial Staff for the Journal format, RStudio for jss_article, James Balamuta for the theme hook, and David Dalpiaz for skeleton content

Examples

if (FALSE) {
library(rmarkdown)
draft("MyArticle.Rmd", template = "latex_journal_format", package = "uiucthemes")
}