Beamer Illinois Market Information Lab Presentation Template for RMarkdown

Generates from an RMarkdown file a Beamer presentation with "marketlab" (UIUC) colors and identity standards with a focus on "Orange" and "White".

beamer_mil(
  toc = FALSE,
  slide_level = 3,
  incremental = FALSE,
  fig_width = 10,
  fig_height = 7,
  fig_crop = TRUE,
  fig_caption = TRUE,
  dev = "pdf",
  df_print = "default",
  fonttheme = "default",
  highlight = "default",
  keep_tex = FALSE,
  latex_engine = "pdflatex",
  citation_package = c("default", "natbib", "biblatex"),
  includes = NULL,
  md_extensions = NULL,
  pandoc_args = NULL
)

Arguments

toc
TRUE to include a table of contents in the output (only level 1 headers will be included in the table of contents).

slide_level
The heading level which defines individual slides. By default this is the highest header level in the hierarchy that is followed immediately by content, and not another header, somewhere in the document. This default can be overridden by specifying an explicit slide_level.

incremental
TRUE to render slide bullets incrementally. Note that if you want to reverse the default incremental behavior for an individual bullet you can precede it with >. For example: > - Bullet Text

fig_width
Default width (in inches) for figures

fig_height
Default height (in inches) for figures

fig_crop
Whether to crop PDF figures with the command pdfcrop. This requires the tools pdfcrop and ghostscript to be installed. By default, fig_crop = TRUE if these two tools are available.

fig_caption
TRUE to render figures with captions

dev
Graphics device to use for figure output (defaults to pdf)

df_print
Method to be used for printing data frames. Valid values include "default", "kable", "tibble", and "paged". The "default" method uses a corresponding S3 method of print, typically print.data.frame. The "kable" method uses the knitr::kable function. The "tibble" method uses the tibble package to print a summary of the data frame. The "paged" method creates a paginated HTML table (note that this method is only valid for formats that produce HTML). In addition to the named methods you can also pass an arbitrary function to be used for printing data frames. You can disable the df_print behavior entirely by setting the option rmarkdown.df_print to FALSE. See Data frame printing section in bookdown book for examples.

fonttheme
Beamer font theme (e.g. "structurebold").

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.

keep_tex
Keep the intermediate tex file used in the conversion to PDF

latex_engine
LaTeX engine for producing PDF output. Options are "pdflatex", "lualatex", "xelatex" and "tectonic".

citation_package
The LaTeX package to process citations, natbib or biblatex. Use default if neither package is to be used, which means citations will be processed via the command pandoc-citeproc.

includes
Named list of additional content to include within the document (typically created using the includes function).

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

pandoc_args
Additional command line options to pass to pandoc

Value

A modified beamer_presentation based on the "marketlab" (UIUC) Beamer themed template.

Author

Jose Luis Rodriguez (Beamer Marketlab Theme) James Joseph Balamuta (Theme Hooks) See AUTHORS for more specific details behind each of the contributions.

Examples

if (FALSE) {
# Generate slide deck from beamer template
rmarkdown::draft("slide_deck.Rmd", template = "beamer_mil", package = "uiucthemes")

# Compile the document
rmarkdown::render("slide_deck/slide_deck.Rmd")
}